requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==1.11.7
django-debug-toolbar==1.9.1
django-nested-admin==3.0.21
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.13
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
msgpack-python==0.5.6
python-monkey-business==1.0.0
pytz==2017.3
pyzmq==17.0.0
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
Werkzeug==0.14.1
